refactor: move types under shared directory & update import paths
